Better install script, added init script

+++ install_dev_tools_debian.sh
@@ -8,31 +8,40 @@ sudo apt-get update
 sudo apt-get install -y tmux git
 
 # install docker
-sudo apt-get install -y \
-    apt-transport-https \
-    ca-certificates \
-    curl \
-    gnupg-agent \
-    software-properties-common
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
-sudo add-apt-repository \
-   "deb [arch=amd64] https://download.docker.com/linux/debian \
-   $(lsb_release -cs) \
-   stable"
-sudo apt-get update
-sudo apt-get install -y docker-ce docker-ce-cli containerd.io
-sudo groupadd docker || true
-sudo usermod -aG docker $USER
-newgrp docker
+if ! command -v docker; then
+    sudo apt-get install -y \
+        apt-transport-https \
+        ca-certificates \
+        curl \
+        gnupg-agent \
+        software-properties-common
+    curl -fsSL https://download.docker.com/linux/debian/gpg | sudo apt-key add -
+    sudo add-apt-repository \
+       "deb [arch=amd64] https://download.docker.com/linux/debian \
+       $(lsb_release -cs) \
+       stable"
+    sudo apt-get update
+    sudo apt-get install -y docker-ce docker-ce-cli containerd.io
+    sudo groupadd docker || true
+    sudo usermod -aG docker $USER
+    newgrp docker || true
+fi
 
-# check the latest version at https://golang.org/dl/
-GO_VERSION=1.15.4
-curl -O https://dl.google.com/go/go${GO_VERSION}.linux-amd64.tar.gz
-tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
-rm -f go${GO_VERSION}.linux-amd64.tar.gz
-sudo rm -rf /usr/local/go
-sudo mv go /usr/local/
+if ! command -v go; then
+    # check the latest version at https://golang.org/dl/
+    GO_VERSION=1.15.4
+    curl -O https://dl.google.com/go/go${GO_VERSION}.linux-amd64.tar.gz
+    tar -xvf go${GO_VERSION}.linux-amd64.tar.gz
+    rm -f go${GO_VERSION}.linux-amd64.tar.gz
+    sudo rm -rf /usr/local/go
+    sudo mv go /usr/local/
 
-mkdir -p $HOME/work
-echo "export GOPATH=\$HOME/work
-export PATH=\$PATH:/usr/local/go/bin:\$GOPATH/bin" > $HOME/.profile
+    mkdir -p $HOME/work
+    if grep -q GOPATH $HOME/.profile; then
+        echo "GOPATH config already exists"
+    else
+        echo "
+export GOPATH=\$HOME/work
+export PATH=\$PATH:/usr/local/go/bin:\$GOPATH/bin" >> $HOME/.profile
+    fi
+fi
